[FIR] Remove misleading code
This code defines "filters", not the
actual dependencies. The actual
dependencies come from the
`binaryModuleData`, but in this code it's
always empty.

The dependencies are actually collected
later. For examole, when FIR/JS builds
the library session, it also calls
`getAllJsDependenciesPaths` to pick up
the same dependencies, and then supplies
them to the klib provider.